+++ gama/GamaClassifier.py
@@ -9,10 +9,11 @@ from sklearn.preprocessing import LabelEncoder
 from .gama import Gama
 from gama.data import X_y_from_arff
 from gama.configuration.classification import clf_config
-from gama.genetic_programming.algorithms.metrics import scoring_to_metric
+from gama.utilities.metrics import scoring_to_metric
 
 
 class GamaClassifier(Gama):
+    """ Wrapper for the toolbox logic executing the AutoML pipeline for (multi-class) classification. """
     def __init__(self, config=None, scoring='neg_log_loss', *args, **kwargs):
         if not config:
             # Do this to avoid the whole dictionary being included in the documentation.
@@ -31,9 +32,15 @@ class GamaClassifier(Gama):
     def _predict(self, x: pd.DataFrame):
         """ Predict the target for input X.
 
-        :param x: a 2d numpy array with the length of the second dimension is equal to that of X of `fit`.
-        :return: a numpy array with predictions. The array is of shape (N,) where N is the length of the
-            first dimension of X.
+        Parameters
+        ----------
+        x: pandas.DataFrame
+            A dataframe with the same number of columns as the input to `fit`.
+
+        Returns
+        -------
+        numpy.ndarray
+            array with predictions of shape (N,) where N is the length of the first dimension of X.
         """
         y = self.model.predict(x)
         # Decode the predicted labels - necessary only if ensemble is not used.
@@ -46,26 +53,40 @@ class GamaClassifier(Gama):
 
         Predict target for x, using the best found pipeline(s) during the `fit` call.
 
-        :param x: a 2d numpy array with the length of the second dimension is equal to that of X of `fit`.
-        :return: a numpy array with class probabilities. The array is of shape (N, K) where N is the length of the
-            first dimension of X, and K is the number of class labels found in `y` of `fit`.
+        Parameters
+        ----------
+        x: pandas.DataFrame
+            A dataframe with the same number of columns as the input to `fit`.
+
+        Returns
+        -------
+        numpy.ndarray
+            Array of shape (N, K) with class probabilities where N is the length of the
+            first dimension of x, and K is the number of class labels found in `y` of `fit`.
         """
         return self.model.predict_proba(x)
 
-    def predict_proba(self, X: Union[pd.DataFrame, np.ndarray]):
-        """ Predict the class probabilities for input X.
+    def predict_proba(self, x: Union[pd.DataFrame, np.ndarray]):
+        """ Predict the class probabilities for input x.
 
-        Predict target for X, using the best found pipeline(s) during the `fit` call.
+        Predict target for x, using the best found pipeline(s) during the `fit` call.
 
-        :param X: a 2d numpy array with the length of the second dimension is equal to that of X of `fit`.
-        :return: a numpy array with class probabilities. The array is of shape (N, K) where N is the length of the
-            first dimension of X, and K is the number of class labels found in `y` of `fit`.
+        Parameters
+        ----------
+        x: pandas.DataFrame or numpy.ndarray
+            A dataframe or numpy array with the same number of columns as the input to `fit`.
+
+        Returns
+        -------
+        numpy.ndarray
+            Array of shape (N, K) with class probabilities where N is the length of the
+            first dimension of x, and K is the number of class labels found in `y` of `fit`.
         """
-        if isinstance(X, np.ndarray):
-            X = pd.DataFrame(X)
+        if isinstance(x, np.ndarray):
+            x = pd.DataFrame(x)
             for col in self._X.columns:
-                X[col] = X[col].astype(self._X[col].dtype)
-        return self._predict_proba(X)
+                x[col] = x[col].astype(self._X[col].dtype)
+        return self._predict_proba(x)
 
     def predict_proba_arff(self, arff_file_path: str):
         """ Predict the class probabilities for input in the arff_file, must have empty target column.
@@ -81,6 +102,7 @@ class GamaClassifier(Gama):
         return self._predict_proba(X)
 
     def fit(self, x, y, *args, **kwargs):
+        """ Should use base class documentation. """
         y_ = y.squeeze() if isinstance(y, pd.DataFrame) else y
         self._label_encoder = LabelEncoder().fit(y_)
         if any([isinstance(yi, str) for yi in y_]):