+++ gama/__init__.py
@@ -1,4 +1,6 @@
-from .GamaClassifier import  GamaClassifier
+from .GamaClassifier import GamaClassifier
 from .GamaRegressor import GamaRegressor
 
 name = "gama"
+
+__all__ = ['GamaClassifier', 'GamaRegressor']