+++ docs/source/releases.rst
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+Version 19.08.0
+---------------
+- Prototype dash app for visualizing GAMA logs.
+- Easy switching between search algorithms
+
 Version 0.1.0
 -------------
 First GAMA release.